Move the autocomplete helpers into the code package, out of the types package.
Doesn't need sharing, and it _does_ need ProviderSettingsManager, so now we don't need to duplicate the type definition
